+++ plugin/src/main/kotlin/de/infolektuell/gradle/jextract/extensions/JextractExtension.kt
@@ -4,6 +4,7 @@ import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Nested
 import javax.inject.Inject
 
@@ -23,6 +24,8 @@ abstract class JextractExtension @Inject constructor(objects: ObjectFactory) {
     }
     /** specify the directory to place generated files of all libraries by default */
     abstract val output: DirectoryProperty
+    /** Generate source files instead of class files for all libraries where not set explicitly (Jextract 21 and below) */
+    abstract val generateSourceFiles: Property<Boolean>
     companion object {
         const val EXTENSION_NAME = "jextract"
     }